- moved function that checks the limits of the TMS to the base tile provider - check limits of TMS in wmts provider
