+++ pygeoapi/provider/base_mvt.py
@@ -249,45 +249,3 @@ class BaseMVTProvider(BaseTileProvider):
             ]
         }
         return links
-
-    def get_tilematrixset(self, tileMatrixSetId):
-        """
-        Get tilematrixset
-
-        :param tileMatrixSetId: tilematrixsetid str
-
-        :returns: tilematrixset enum object
-        """
-
-        enums = [e.value for e in TileMatrixSetEnum]
-        enum = None
-
-        try:
-            for e in enums:
-                if tileMatrixSetId == e.tileMatrixSet:
-                    enum = e
-            if not enum:
-                raise ValueError('could not find this tilematrixset')
-            return enum
-
-        except ValueError as err:
-            LOGGER.error(err)
-
-    def is_in_limits(self, tilematrixset, z, x, y):
-        """
-        Is within the limits of the tilematrixset
-
-        :param z: tilematrix
-        :param x: x
-        :param y: y
-
-        :returns: wether this tile is within the tile matrix
-        set limits (Boolean)
-        """
-
-        try:
-            if int(x) < tilematrixset.tileMatrices[int(z)]['matrixWidth'] and int(y) < tilematrixset.tileMatrices[int(z)]['matrixHeight']: # noqa
-                return True
-            return False
-        except ValueError as err:
-            LOGGER.error(err)
